Move RIBS store implementation to a separate module
To avoid dependency creep, move RIBS store into a separate go module
under `integration`.

Introduce go workspace to assure all modules remain compilable.

Fix docker build

Build RIBS integration in a separate docker layer since it does not
change often.
