+++ .github/workflows/container-build.yml
@@ -15,4 +15,6 @@ jobs:
         uses: docker/build-push-action@v3
         with:
           context: .
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
           push: false
