+++ Dockerfile
@@ -6,10 +6,9 @@ RUN apt-get update && apt-get install -y libhwloc-dev ocl-icd-opencl-dev jq
 COPY Makefile .
 RUN make extern/filecoin-ffi
 
-COPY go.* .
+COPY . .
 RUN go mod download
 
-COPY . .
 RUN CGO_ENABLED=1 go build -o /go/bin/motion ./cmd/motion
 
 FROM gcr.io/distroless/base-debian11
