+++ cmd/motion/main.go
@@ -38,11 +38,6 @@ func main() {
 				Value:       os.TempDir(),
 				EnvVars:     []string{"MOTION_STORE_DIR"},
 			},
-			&cli.BoolFlag{
-				Name:        "experimentalRibsStore",
-				Usage:       "Whether to use experimental RIBS as the storage and deal making",
-				DefaultText: "Local storage is used",
-			},
 			&cli.StringFlag{
 				Name:        "localWalletDir",
 				Usage:       "The path to the local wallet directory.",
@@ -106,18 +101,7 @@ func main() {
 			var store blob.Store
 			lotusAPI := cctx.String("lotusApi")
 			lotusToken := cctx.String("lotusToken")
-			if cctx.Bool("experimentalRibsStore") {
-				rbstore, err := blob.NewRibsStore(storeDir, ks)
-				if err != nil {
-					return err
-				}
-				logger.Infow("Using RIBS blob store", "storeDir", storeDir)
-				if err := rbstore.Start(cctx.Context); err != nil {
-					logger.Errorw("Failed to start RIBS blob store", "err", err)
-					return err
-				}
-				store = rbstore
-			} else if cctx.Bool("experimentalSingularityStore") {
+			if cctx.Bool("experimentalSingularityStore") {
 				singularityAPIUrl := cctx.String("experimentalRemoteSingularityAPIUrl")
 				var client client.Client
 				if singularityAPIUrl != "" {
@@ -182,12 +166,6 @@ func main() {
 			if err := m.Shutdown(ctx); err != nil {
 				logger.Warnw("Failure occurred while shutting down Motion.", "err", err)
 			}
-			// TODO: Re-enable once the panic is fixed. See: https://github.com/FILCAT/ribs/issues/39
-			//if rbstore, ok := store.(*blob.RibsStore); ok {
-			//	if err := rbstore.Shutdown(ctx); err != nil {
-			//		logger.Warnw("Failure occurred while shutting down RIBS blob store.", "err", err)
-			//	}
-			//}
 			logger.Info("Shut down Motion successfully.")
 			return nil
 		},
