+++ go.work
@@ -1,0 +1,6 @@
+go 1.20
+
+use (
+	.
+	integration/ribs
+)
