+++ integration/ribs/README.md
@@ -1,0 +1,3 @@
+# Experimental RIBS Motion-store
+
+This module offers an experimental [RIBS](https://github.com/FILCAT/ribs) Motion-compatible store implementations.
